+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - A ban reason can now be provided when banning a user from this screen.
   - Badges are now displayed in this dialog.
 - Actions can now be reordered.
+- Youtube video previews now indicate the video duration.
 - Add a button & context menu item to unban previously banned users.
 - Add a button to unblock previously blocked users.